+++ pyweather.py
@@ -47,7 +47,7 @@ config.read('storage//config.ini')
 # because it's not created. Creative.
 try:
     configprovisioned = config.getboolean('USER', 'configprovisioned')
-except KeyError:
+except:
     print("We ran into an error. Full traceback below.")
     traceback.print_exc()
     # Not doing a separate script launch. Doing sys.exit in that script would just continue